# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TruckGo API - Uber-for-trucks platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CustomerRegistrationInput {
  name: string;
  email: string;
  phone: string;
  password: string;
}

export interface DriverRegistrationInput {
  name: string;
  email: string;
  phone: string;
  password: string;
  licenseNumber: string;
  truckTypeId: number;
  vehiclePlate: string;
  /** @nullable */
  vehicleYear?: number | null;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  customer: 'customer',
  driver: 'driver',
} as const;

export interface LoginInput {
  email: string;
  password: string;
  role: LoginInputRole;
}

export type AuthResponseRole = typeof AuthResponseRole[keyof typeof AuthResponseRole];


export const AuthResponseRole = {
  customer: 'customer',
  driver: 'driver',
} as const;

export interface AuthResponse {
  userId: number;
  role: AuthResponseRole;
  name: string;
  email: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  customer: 'customer',
  driver: 'driver',
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserProfileRole;
  createdAt?: string;
}

export interface TruckType {
  id: number;
  name: string;
  description: string;
  capacityKg: number;
  basePrice: number;
  pricePerKm: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface EstimateInput {
  truckTypeId: number;
  pickupAddress: string;
  dropoffAddress: string;
  distanceKm: number;
}

export interface EstimateResult {
  truckTypeId: number;
  distanceKm: number;
  basePrice: number;
  pricePerKm: number;
  estimatedTotal: number;
  estimatedMinutes: number;
}

export interface BookingInput {
  customerId: number;
  truckTypeId: number;
  pickupAddress: string;
  dropoffAddress: string;
  distanceKm: number;
  estimatedPrice: number;
  goodsDescription: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  driverId?: number | null;
  /** @nullable */
  driverName?: string | null;
  /** @nullable */
  driverPhone?: string | null;
  truckTypeId: number;
  /** @nullable */
  truckTypeName?: string | null;
  pickupAddress: string;
  dropoffAddress: string;
  distanceKm: number;
  estimatedPrice: number;
  /** @nullable */
  finalPrice?: number | null;
  status: BookingStatus;
  goodsDescription: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface BookingSummary {
  total: number;
  pending: number;
  active: number;
  completed: number;
  cancelled: number;
  totalSpent: number;
}

export interface AcceptBookingInput {
  driverId: number;
}

export type DriverStatus = typeof DriverStatus[keyof typeof DriverStatus];


export const DriverStatus = {
  available: 'available',
  busy: 'busy',
  offline: 'offline',
} as const;

export interface Driver {
  id: number;
  name: string;
  phone: string;
  licenseNumber: string;
  vehiclePlate: string;
  /** @nullable */
  vehicleYear?: number | null;
  truckTypeId: number;
  /** @nullable */
  truckTypeName?: string | null;
  status: DriverStatus;
  rating: number;
  totalTrips: number;
  createdAt?: string;
}

export type DriverStatusUpdateStatus = typeof DriverStatusUpdateStatus[keyof typeof DriverStatusUpdateStatus];


export const DriverStatusUpdateStatus = {
  available: 'available',
  busy: 'busy',
  offline: 'offline',
} as const;

export interface DriverStatusUpdate {
  status: DriverStatusUpdateStatus;
}

export interface DriverStats {
  driverId: number;
  totalTrips: number;
  completedTrips: number;
  cancelledTrips: number;
  totalEarnings: number;
  rating: number;
  reviewCount: number;
}

export interface ReviewInput {
  bookingId: number;
  customerId: number;
  driverId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
}

export interface Review {
  id: number;
  bookingId: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  driverId: number;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface AdminStats {
  totalBookings: number;
  pendingBookings: number;
  activeBookings: number;
  completedBookings: number;
  cancelledBookings: number;
  totalRevenue: number;
  totalDrivers: number;
  availableDrivers: number;
  busyDrivers: number;
  offlineDrivers: number;
  totalCustomers: number;
}

export interface AdminCustomer {
  id: number;
  name: string;
  email: string;
  phone: string;
  createdAt: string;
  totalBookings: number;
  totalSpent: number;
}

export type GetMeParams = {
userId: number;
role: GetMeRole;
};

export type GetMeRole = typeof GetMeRole[keyof typeof GetMeRole];


export const GetMeRole = {
  customer: 'customer',
  driver: 'driver',
} as const;

export type ListBookingsParams = {
customerId: number;
status?: ListBookingsStatus;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type GetBookingSummaryParams = {
customerId: number;
};

export type ListAvailableDriversParams = {
truckTypeId?: number;
};

export type ListAdminBookingsParams = {
status?: ListAdminBookingsStatus;
};

export type ListAdminBookingsStatus = typeof ListAdminBookingsStatus[keyof typeof ListAdminBookingsStatus];


export const ListAdminBookingsStatus = {
  pending: 'pending',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;