# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TruckGo API - Uber-for-trucks platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/customer:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistrationInput"
      responses:
        "201":
          description: Customer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/register/driver:
    post:
      operationId: registerDriver
      tags: [auth]
      summary: Register a new driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverRegistrationInput"
      responses:
        "201":
          description: Driver registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (customer or driver)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
        - in: query
          name: role
          required: true
          schema:
            type: string
            enum: [customer, driver]
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # --- Truck Types ---
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for the current customer
      parameters:
        - in: query
          name: customerId
          required: true
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, cancelled]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error
  /bookings/summary:
    get:
      operationId: getBookingSummary
      tags: [bookings]
      summary: Get booking summary stats for a customer (dashboard)
      parameters:
        - in: query
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingSummary"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
  /bookings/{id}/cancel:
    patch:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found

  # --- Drivers ---
...
  /bookings/{id}/accept:
    patch:
      operationId: acceptBooking
      tags: [bookings]
      summary: Driver accepts a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptBookingInput"
      responses:
        "200":
          description: Booking accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/start:
    patch:
      operationId: startBooking
      tags: [bookings]
      summary: Driver marks pickup complete and trip started
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/complete:
    patch:
      operationId: completeBooking
      tags: [bookings]
      summary: Driver marks delivery complete
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # --- Reviews ---
...
components:
  schemas:
...
    CustomerRegistrationInput:
      type: object
      required: [name, email, phone, password]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
    DriverRegistrationInput:
      type: object
      required: [name, email, phone, password, licenseNumber, truckTypeId, vehiclePlate]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        licenseNumber:
          type: string
        truckTypeId:
          type: integer
        vehiclePlate:
          type: string
        vehicleYear:
          type: ["integer", "null"]
    LoginInput:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, driver]
    AuthResponse:
      type: object
      required: [userId, role, name, email]
      properties:
        userId:
          type: integer
        role:
          type: string
          enum: [customer, driver]
        name:
          type: string
        email:
          type: string
    UserProfile:
      type: object
      required: [id, name, email, phone, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, driver]
        createdAt:
          type: string
          format: date-time

    # --- Truck Types ---
...
    BookingInput:
      type: object
      required: [customerId, truckTypeId, pickupAddress, dropoffAddress, distanceKm, estimatedPrice, goodsDescription]
      properties:
        customerId:
          type: integer
        truckTypeId:
          type: integer
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        distanceKm:
          type: number
        estimatedPrice:
          type: number
        goodsDescription:
          type: string
        notes:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
    Booking:
      type: object
      required: [id, customerId, truckTypeId, pickupAddress, dropoffAddress, distanceKm, estimatedPrice, status, goodsDescription, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        driverId:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
        driverPhone:
          type: ["string", "null"]
        truckTypeId:
          type: integer
        truckTypeName:
          type: ["string", "null"]
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        distanceKm:
          type: number
        estimatedPrice:
          type: number
        finalPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        goodsDescription:
          type: string
        notes:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    BookingSummary:
      type: object
      required: [total, pending, active, completed, cancelled, totalSpent]
      properties:
        total:
          type: integer
        pending:
          type: integer
        active:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
        totalSpent:
          type: number
    AcceptBookingInput:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: integer

    # --- Drivers ---
...