# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TruckGo API - Uber-for-trucks platform
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  driverId?: number | null;
  /** @nullable */
  driverName?: string | null;
  /** @nullable */
  driverPhone?: string | null;
  truckTypeId: number;
  /** @nullable */
  truckTypeName?: string | null;
  pickupAddress: string;
  dropoffAddress: string;
  distanceKm: number;
  estimatedPrice: number;
  /** @nullable */
  finalPrice?: number | null;
  status: BookingStatus;
  goodsDescription: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  scheduledAt?: Date | null;
  /** @nullable */
  startedAt?: Date | null;
  /** @nullable */
  completedAt?: Date | null;
  createdAt: Date;
}