# lib/api-zod/src/generated/types/customerRegistrationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * TruckGo API - Uber-for-trucks platform
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerRegistrationInput {
  name: string;
  email: string;
  phone: string;
  password: string;
}